# examples/openapi/srcgen/openapi.yaml
openapi: 3.1.0
info:
  title: Test API
  description: |
    This example should test every option of this generator.
  version: '1.0'
servers:
  - url: '/api/v1'
security:
  - ApiKeyAuth: []
...
paths:  
...
  '/users':
    get:
      tags:
        - Management
      summary: List Users
      operationId: getUsers
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/count'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        default:
          $ref: '#/components/responses/DefaultError'
    post:
      tags:
        - Management
      summary: Create a User
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          $ref: '#/components/responses/DefaultError'
  '/users/{id}':
    get:
      tags:
        - Management
      summary: Read the specified User
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          $ref: '#/components/responses/DefaultError'
    put:
      tags:
        - Management
      summary: Replace the specified User
      operationId: replaceUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          $ref: '#/components/responses/DefaultError'
    patch:
      tags:
        - Management
      summary: Update the specified User
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          $ref: '#/components/responses/DefaultError'
    delete:
      tags:
        - Management
      summary: Delete the specified User
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        default:
          $ref: '#/components/responses/DefaultError'
  '/projects/{project_id}/tickets':
    get:
      tags:
        - Incident
      summary: List Tickets
      operationId: getTickets
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/count'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketList'
        default:
          $ref: '#/components/responses/DefaultError'
    post:
      tags:
        - Incident
      summary: Create a Ticket
      operationId: createTicket
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ticket'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        default:
          $ref: '#/components/responses/DefaultError'
  '/projects/{project_id}/tickets/{id}':
    get:
      tags:
        - Incident
      summary: Read the specified Ticket
      operationId: getTicket
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        default:
          $ref: '#/components/responses/DefaultError'
    put:
      tags:
        - Incident
      summary: Replace the specified Ticket
      operationId: replaceTicket
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ticket'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        default:
          $ref: '#/components/responses/DefaultError'
    patch:
      tags:
        - Incident
      summary: Update the specified Ticket
      operationId: updateTicket
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ticket'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        default:
          $ref: '#/components/responses/DefaultError'
    delete:
      tags:
        - Incident
      summary: Delete the specified Ticket
      operationId: deleteTicket
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        default:
          $ref: '#/components/responses/DefaultError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  responses:
    DefaultError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppError'
  parameters:
    cursor:
      name: cursor
      description: |
        Paginate through collections of data by setting the cursor parameter to a next_cursor attribute returned by a previous response.
      in: query
      required: false
      schema:
        type: string
    limit:  
      name: limit
      description: |
        The maximum number of items to return.
      in: query
      required: false
      schema:
        type: integer
        format: int64
    count:
      name: count
      description: |
        Returns the total number of items in the collection.
      in: query
      required: false
      schema:
        type: boolean
  schemas:
    Status:
      type: string
      enum:
        - OPEN
        - TO_DO
        - CLOSED
...
    UserList:
      required:
        - items
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        next_cursor:
          type: string
          maxLength: 255
        total_count:
          type: integer
          format: int64
          minimum: 0
    User:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          minimum: 0
        name:
          type: integer
          format: int32
          minimum: 0
        full_name:
          type: string
          maxLength: 255
        emails:
          type: array
          items:
            type: string
            maxLength: 255
        password:
          type: string
          maxLength: 255
    TicketList:
      required:
        - items
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Ticket'
        next_cursor:
          type: string
          maxLength: 255
        total_count:
          type: integer
          format: int64
          minimum: 0
    Ticket:
      required:
        - id
        - number
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          minimum: 0
        number:
          type: integer
          format: int32
          minimum: 0
        title:
          type: string
          maxLength: 255
          default: Duke
        issuer_id:
          type: integer
          format: int32
          minimum: 0
          default: 42
        issuer_body:
          $ref: '#/components/schemas/User'
          readOnly: true
        body:
          type: string
          maxLength: 255
        status:
          $ref: '#/components/schemas/Status'
          default: OPEN
...
    AppError:
      required:
        - code
        - message
      type: object
      properties:
        code:
          description: A identifier that categorizes the error
          type: string
        message:
          description: A brief, human-readable message about the error
          type: string
        status:
          description: The HTTP response code
          type: integer
          format: int32
        path:
          description: A URI that identifies the specific occurrence of the error
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          description: Detailed explanations of the error
          type: array
          items:
            type: string