# examples/spring/srcgen/src/main/java/app/server/controller/TicketController.java
/**
 * NOTE: This class was generated by QSDL.
 */

package app.server.controller;

import app.server.api.TicketApi;
import app.server.domain.*;
import app.server.model.*;
import app.server.service.TicketService;
import app.server.util.Validator;

import lombok.AllArgsConstructor;
import org.springframework.http.*;
import org.springframework.stereotype.Controller;


@Controller
@AllArgsConstructor
public class TicketController extends BaseController implements TicketApi {

  private final TicketService ticketService;

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<CursorPage<Ticket>> getTickets(CursorPageable pageable) {
    CursorPage<Ticket> response = ticketService.getTickets(pageable, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<Ticket> createTicket(Ticket body) {
    Validator.validate(body);
    Ticket response = ticketService.createTicket(body, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<Ticket> getTicket(Long id) {
    Ticket response = ticketService.getTicket(id, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<Ticket> replaceTicket(Long id, Ticket body) {
    Validator.validate(body);
    Ticket response = ticketService.replaceTicket(id, body, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<Ticket> updateTicket(Long id, Ticket body) {
    Validator.validateExRequired(body);
    Ticket response = ticketService.updateTicket(id, body, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<Void> deleteTicket(Long id) {
    ticketService.deleteTicket(id, super.getContext());
    return new ResponseEntity<>(HttpStatus.OK);
  }

}