# examples/spring/srcgen/src/main/java/app/server/controller/UserController.java
/**
 * NOTE: This class was generated by QSDL.
 */

package app.server.controller;

import app.server.api.UserApi;
import app.server.domain.*;
import app.server.model.*;
import app.server.service.UserService;
import app.server.util.Validator;

import lombok.AllArgsConstructor;
import org.springframework.http.*;
import org.springframework.stereotype.Controller;


@Controller
@AllArgsConstructor
public class UserController extends BaseController implements UserApi {

  private final UserService userService;

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<CursorPage<User>> getUsersForTicket(Long ticketId, CursorPageable pageable) {
    CursorPage<User> response = userService.getUsersForTicket(ticketId, pageable, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<Void> addUserToTicket(Long ticketId, Long id) {
    userService.addUserToTicket(ticketId, id, super.getContext());
    return new ResponseEntity<>(HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<Void> removeUserFromTicket(Long ticketId, Long id) {
    userService.removeUserFromTicket(ticketId, id, super.getContext());
    return new ResponseEntity<>(HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<CursorPage<User>> getUsers(CursorPageable pageable) {
    CursorPage<User> response = userService.getUsers(pageable, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<User> createUser(User body) {
    Validator.validate(body);
    User response = userService.createUser(body, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<User> getUser(Long id) {
    User response = userService.getUser(id, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<User> replaceUser(Long id, User body) {
    Validator.validate(body);
    User response = userService.replaceUser(id, body, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<User> updateUser(Long id, User body) {
    Validator.validateExRequired(body);
    User response = userService.updateUser(id, body, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<Void> deleteUser(Long id) {
    userService.deleteUser(id, super.getContext());
    return new ResponseEntity<>(HttpStatus.OK);
  }

}